feat: 경로 생성 & 조회 기능 구현

+++ src/main/java/zeph_server/course/dto/common/Point.java
@@ -2,7 +2,7 @@ package zeph_server.course.dto.common;
 
 public record Point(
         Long id,
-        double lat,
-        double lng
+        Float lat,
+        Float lng
 ) {
 }
